+++ geolib/__init__.py
@@ -3,6 +3,3 @@ GEOLib Library
 """
 
 __version__ = "2.3.0"
-
-from . import utils
-from .models import *
