+++ tests/service/test_service.py
@@ -11,7 +11,7 @@ import pytest
 from fastapi.testclient import TestClient
 from requests.auth import HTTPBasicAuth
 
-from geolib import BaseModelList, DFoundationsModel, DSettlementModel
+from geolib.models import BaseModelList, DFoundationsModel, DSettlementModel
 from geolib.service.main import app
 from tests.utils import TestUtils, only_teamcity
